requirements.txt
boto3==1.26.12
krakenapi==1.0.0a7
PyYAML==6.0